Added sse_matrix.cc file, added pinhole camera implementation

+++ core/camera.h
@@ -41,7 +41,7 @@ namespace pixel {
     class camera {
     public:
         // Create ray for a given couple of pixel coordinates and a sample
-        virtual ray generate_ray(const uint32_t i, const uint32_t j, const double u1, const double u2) const = 0;
+        virtual ray generate_ray(const uint32_t i, const uint32_t j, const float u1, const float u2) const = 0;
     };
 }
